requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-chroma
python-dotenv
pypdf
